(Incomplete) Found a design mistake
There's no joinable field between Request and Pre-prepare message.
Because, a replica need to find a timestamp and an operation, which are inside a request message at a commit phase.
It is easy to find out pre-prepare message from a commit message because they share the sequence number.
But, a request message doesn't have it.

So, the only thing request and pre-prepare message commonly have is the operation.
But, we can't guarantee that each operation is unique.
Therefore, we need to insert some random thing inside the operaion to guarantee uniqueness.
Also, a client must not reuse any operation instance!